Implementing Asset Registry
