+++ src/public/Base/GameInstance.h
@@ -1,7 +1,5 @@
 #pragma once
 #include "Base/Core.h"
-#include <set>
-#include <sstream>
 #include <algorithm>
 #include "Base/EventDispatcher.hpp"
 #include "Base/ResourceManager.hpp"
@@ -9,6 +7,7 @@
 #include "StateMachines/GameModeSwitcher.h"
 #include "Base/GameThreadQueue.hpp"
 #include "Base/GameMode.h"
+#include "Base/AssetRegistry.hpp"
 
 struct WindowProperties
 {
@@ -97,8 +96,6 @@ private:
 	~GameInstance() = default;
 
 	static GameInstance* g_Instance;
-	static std::unordered_map<std::string, std::set<int32_t>> g_AssetRegistry;
-
 
 protected:
 
@@ -116,12 +113,12 @@ protected:
 	*****************************
 	*/
 
-	static ResourceManager m_ResourceManager;
+	static ResourceManager g_ResourceManager;
+	static AssetRegistry g_AssetRegistry;
 
 	static void CreateWindow();
 	static void GameLoop();
 
-	bool ParseAssetName(const std::string& FullName, std::string& OutBaseName, int32_t& OutNumber);
 
 public:
 
@@ -136,56 +133,6 @@ public:
 
 	TextureResource* GetResource(std::string Name);
 
-	/*
-	*****************************
-	* Asset Registry
-	* INFO:
-	* This could (and should) be refactored to its own Class
-	* also this could reside inside the GameMode, cause in the current Implemention, only one GameMode can be active at a time
-	* but for future use or expansion, i will let them seperate for now
-	* but it still should be its own class
-	* 
-	*****************************
-	*/
-
-	template <typename T>
-	std::shared_ptr<T> LoadAssetFromSoftObjectPath(SoftObjectPath<T> Path)
-	{
-		std::string FullPath = Path.ToString();
-		size_t Index = FullPath.find_first_of("/");
-
-		std::string GameMode = FullPath.substr(0, Index);
-		std::string Object = FullPath.substr(Index + 1);
-
-		if (g_ActiveStateMachine.GetCurrentGameMode()->GetName() != GameMode)
-		{
-			return nullptr;
-		}
-
-		auto MapIT = g_ActiveStateMachine.GetCurrentGameMode()->m_Objects.find(FullPath);
-
-		if (MapIT == g_ActiveStateMachine.GetCurrentGameMode()->m_Objects.end())
-		{
-			return nullptr;
-		}
-
-		std::shared_ptr<T> CastedOBJ = std::dynamic_pointer_cast<T>(MapIT->second);
-
-		return CastedOBJ;
-	}
-
-	// Static helper function
-	template <typename T>
-	static std::shared_ptr<T> LoadFromSoftObjectPath(const SoftObjectPath<T>& path)
-	{
-		return GetInstance()->LoadAssetFromSoftObjectPath(path);
-	}
-
-	std::string RegisterAsset(const std::string name);
-
-	bool UnregisterAsset(const std::string name);
-
-	std::string GenerateNextAvaiableName(const std::string base_name);
 
 	/*
 	*****************************
@@ -203,6 +150,7 @@ public:
 	WindowProperties GetWindowProperties() const { return m_WindowProperties; }
 
 	static GameMode* GetCurrentGameMode();
+	static AssetRegistry* GetAssetRegistry();
 
 
 	static EventDispatcher& GetUIEventDispatcher();
